Add user and role entities according to silex standard Configure silex firewall based on Basics http auth Add monolog logger support

+++ app/Db/Proxies/__CG__BookmarkerDbEntitiesRole.php
@@ -179,7 +179,7 @@ class Role extends \Bookmarker\Db\Entities\Role implements \Doctrine\ORM\Proxy\P
     public function getId()
     {
         if ($this->__isInitialized__ === false) {
-            return (int)  parent::getId();
+            return  parent::getId();
         }
